+++ lib/playx_theme.dart
@@ -7,7 +7,5 @@ export 'src/model/x_color_scheme.dart';
 export 'src/model/x_theme.dart';
 export 'src/widgets/image_theme.dart';
 export 'src/utils/utils.dart';
-
-
 export 'package:playx_core/playx_core.dart';
 export 'package:flex_seed_scheme/flex_seed_scheme.dart';